mysqlctld: review changes, code refactor, removed some functions.
