+++ go/test/endtoend/backup/vtbackup/backup_only_test.go
@@ -17,17 +17,14 @@ limitations under the License.
 package vtbackup
 
 import (
-	"bufio"
 	"fmt"
-	"os"
-	"path"
-	"strings"
 	"testing"
 	"time"
 
 	"vitess.io/vitess/go/test/endtoend/cluster"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	"vitess.io/vitess/go/vt/log"
 )
@@ -54,9 +51,7 @@ func TestTabletInitialBackup(t *testing.T) {
 	//    - list the backups, remove them
 
 	vtBackup(t, true)
-	backups := countBackups(t)
-	assert.Equal(t, 1, backups)
-
+	localCluster.VerifyBackupCount(t, shardKsName, 1)
 	// Initialize the tablets
 	initTablets(t, false, false)
 
@@ -107,10 +102,11 @@ func firstBackupTest(t *testing.T, tabletType string) {
 	//    - list the backup, remove it
 
 	// Store initial backup counts
-	backupsCount := countBackups(t)
+	listbackups, err := localCluster.ListBackups(shardKsName)
+	require.Nil(t, err)
 
 	// insert data on master, wait for slave to get it
-	_, err := master.VttabletProcess.QueryTablet(vtInsertTest, keyspaceName, true)
+	_, err = master.VttabletProcess.QueryTablet(vtInsertTest, keyspaceName, true)
 	assert.Nil(t, err)
 	// Add a single row with value 'test1' to the master tablet
 	_, err = master.VttabletProcess.QueryTablet("insert into vt_insert_test (msg) values ('test1')", keyspaceName, true)
@@ -125,8 +121,7 @@ func firstBackupTest(t *testing.T, tabletType string) {
 	log.Info("done taking backup %s", time.Now())
 
 	// check that the backup shows up in the listing
-	backups := countBackups(t)
-	assert.Equal(t, backups, backupsCount+1)
+	localCluster.VerifyBackupCount(t, shardKsName, len(listbackups)+1)
 
 	// insert more data on the master
 	_, err = master.VttabletProcess.QueryTablet("insert into vt_insert_test (msg) values ('test2')", keyspaceName, true)
@@ -154,9 +149,8 @@ func firstBackupTest(t *testing.T, tabletType string) {
 		assert.Equal(t, "must_not", result.Rows[3][1].ToString(), "PromotionRule")
 	}
 
-	removeBackups(t)
-	backups = countBackups(t)
-	assert.Equal(t, 0, backups)
+	localCluster.RemoveAllBackups(t, shardKsName)
+	localCluster.VerifyBackupCount(t, shardKsName, 0)
 
 }
 
@@ -166,52 +160,6 @@ func vtBackup(t *testing.T, initialBackup bool) {
 	log.Info("starting backup tablet %s", time.Now())
 	err := localCluster.StartVtbackup(newInitDBFile, initialBackup, keyspaceName, shardName, cell, extraArgs...)
 	assert.Nil(t, err)
-}
-
-func listBackups(t *testing.T) string {
-	// Get a list of backup names for the current shard.
-	localCluster.VtctlProcess = *cluster.VtctlProcessInstance(localCluster.TopoPort, localCluster.Hostname)
-	backups, err := localCluster.VtctlProcess.ExecuteCommandWithOutput(
-		"-backup_storage_implementation", "file",
-		"-file_backup_storage_root",
-		path.Join(os.Getenv("VTDATAROOT"), "tmp", "backupstorage"),
-		"ListBackups", shardKsName,
-	)
-	assert.Nil(t, err)
-	return backups
-}
-
-func countBackups(t *testing.T) int {
-	// Count the number of backups available in current shard.
-	backupList := listBackups(t)
-	backupCount := 0
-	// Counts the available backups
-	scanner := bufio.NewScanner(strings.NewReader(backupList))
-	for scanner.Scan() {
-		if scanner.Text() != "" {
-			backupCount++
-		}
-	}
-	return backupCount
-}
-
-func removeBackups(t *testing.T) {
-	// Remove all the backups from the shard
-	backupList := listBackups(t)
-
-	scanner := bufio.NewScanner(strings.NewReader(backupList))
-	for scanner.Scan() {
-		if scanner.Text() != "" {
-			_, err := localCluster.VtctlProcess.ExecuteCommandWithOutput(
-				"-backup_storage_implementation", "file",
-				"-file_backup_storage_root",
-				path.Join(os.Getenv("VTDATAROOT"), "tmp", "backupstorage"),
-				"RemoveBackup", shardKsName, scanner.Text(),
-			)
-			assert.Nil(t, err)
-		}
-	}
-
 }
 
 func initTablets(t *testing.T, startTablet bool, initShardMaster bool) {
